konvoy-engine: codegen framework — generator-agnostic CodeGenerator trait + cache-key hashing (#289)
Introduce the code-generation framework, decoupled from any concrete generator so
OpenAPI/Fabrikt (next) and future tools (gRPC, …) plug in without touching it.

- CodeGenerator trait: name/display_name/managed_tool/config_hash_parts/input_files/
  generate — the contract every generator implements.
- Deterministic input hashing (compute_codegen_hashes / _hash_pairs /
  compute_generator_hash) folding each generator's config + every input file's path
  and content into a stable, platform-independent cache key. Operates on
  &[Box<dyn CodeGenerator>] — the framework never names a concrete generator; the
  config -> generators registry lives with the implementations.
- The framework owns determinism: inputs are normalized (project-relative, `.`
  collapsed), sorted, and de-duplicated before hashing; non-relative / `..`-escaping
  inputs are rejected; `display_name` is excluded from the key.
- generator_summaries / managed_tools / generator_output_dir helpers.
- error: CodegenInputNotFound (raised when a declared input is missing).
- konvoy-util: sha256_multi length-prefixes parts with a fixed 8-byte width instead
  of `usize`, so the cache key is identical across architectures (byte-for-byte
  unchanged on 64-bit). The framework's portability claim depends on this.
- Tested with a fake generator (zero dependency on any real generator): determinism,
  order/dup/spelling-insensitivity, missing-input error, escaping-input rejection,
  display_name exclusion, and golden digests pinning the v1 wire format (single- and
  multi-file). 22 tests.

First of the 2.2 split (was the monolithic codegen PR). The konvoy-util walk helper
and the OpenAPI/Fabrikt implementation follow, stacked on this.

+++ crates/konvoy-util/src/hash.rs
@@ -104,7 +104,10 @@ pub fn sha256_multi(parts: &[&str]) -> String {
     let mut hasher = Sha256::new();
     for part in parts {
         // Length-prefix each part to avoid collisions like ["ab","c"] vs ["a","bc"].
-        let len_bytes = part.len().to_le_bytes();
+        // Use a fixed 8-byte width (not `usize`, which is 4 bytes on 32-bit and 8 on
+        // 64-bit) so the digest — and any cache key built on it — is identical across
+        // architectures. On 64-bit this is byte-for-byte unchanged.
+        let len_bytes = (part.len() as u64).to_le_bytes();
         hasher.update(len_bytes);
         hasher.update(part.as_bytes());
     }
